refactor(inference): move attention trace diagnostics
